Create timer that ignores inactivity
This is required when the user comes back to the page after inactivity -
without ignoring that timeframe, the graphic breaks.
We detect inactivity windows by measuring a mean duration between each
call to the time function, and if the distance between two calls is
greater than the mean we remove the difference from this and all the
following timestamps.